requirements.txt
asgiref==3.4.1
autopep8==1.6.0
awscli==1.21.10
botocore==1.22.10
colorama==0.4.3
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
docutils==0.15.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
